requirements.txt
mysql-connector-python==8.0.26; sys_platform == "darwin"
mysql-connector-python-rf==2.2.2; sys_platform == "linux"

configparser~=5.0.2
mysql-connector-python~=8.0.26